Fix HostTests compiler warnings

+++ tests/HostTests/modules/Stream.cpp
@@ -36,7 +36,7 @@ public:
 			String s(FS_abstract);
 			MemoryDataStream stream(std::move(s));
 			TEST_ASSERT(!s);
-			TEST_ASSERT(FS_abstract.length() == size_t(stream.available()));
+			TEST_ASSERT((int)FS_abstract.length() == stream.available());
 			// And back again
 			stream.moveString(s);
 			TEST_ASSERT(FS_abstract == s);
@@ -46,7 +46,7 @@ public:
 		TEST_CASE("LimitedMemoryStream::moveString (1)")
 		{
 			FSTR::Stream src(FS_abstract);
-			REQUIRE(size_t(src.available()) == FS_abstract.length());
+			REQUIRE(src.available() == (int)FS_abstract.length());
 			LimitedMemoryStream dest(FS_abstract.length());
 			// Operation will drop last character as this is converted to NUL
 			dest.copyFrom(&src);
@@ -62,7 +62,7 @@ public:
 		TEST_CASE("LimitedMemoryStream::moveString (2)")
 		{
 			FSTR::Stream src(FS_abstract);
-			REQUIRE(size_t(src.available()) == FS_abstract.length());
+			REQUIRE(src.available() == (int)FS_abstract.length());
 			LimitedMemoryStream dest(FS_abstract.length() + 10);
 			// Operation will drop last character as this is converted to NUL
 			dest.copyFrom(&src);
